fix: change elements and code after review
